Extract logic to organize repos, services
