+++ src/lib/services/library/_internal/adapters/web/entities/LibraryAuthor.ts
@@ -1,0 +1,13 @@
+import { LibraryAuthor } from "@/bindings";
+
+import { Author } from "./Author";
+
+export const libraryAuthor = {
+	fromAuthor: (author: Author): LibraryAuthor => {
+		return {
+			id: author.id.toString(),
+			name: author.name,
+			sortable_name: "",
+		};
+	},
+};
