+++ src/lib/services/library/_internal/adapters/web/services/author.ts
@@ -1,0 +1,10 @@
+import { AuthorRepository } from "../repositories/Repository";
+import { Author } from "../entities/Author";
+
+export const createAuthorService = (authorRepository: AuthorRepository) => {
+	return {
+		all: async (): Promise<Author[]> => {
+			return await authorRepository.all();
+		},
+	};
+};
