+++ src/lib/services/library/_internal/adapters/web/types.ts
@@ -1,0 +1,2 @@
+export type AuthorId = string & { readonly __brand: "AuthorId" };
+export type BookId = string & { readonly __brand: "BookId" };
